requirements.txt
fastapi
uvicorn
paho-mqtt
scikit-learn
pandas
numpy
mysql-connector-python
requests